+++ Healenium.Selenium.Tests/Healenium.Selenium.Tests.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 25.0.1700.6
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Healenium.Selenium.Tests", "Healenium.Selenium.Tests\Healenium.Selenium.Tests.csproj", "{313AEA56-2B75-4937-B67B-9979D79428C9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Healenium.Selenium", "Healenium.Selenium\Healenium.Selenium.csproj", "{4629A941-429D-4843-BA72-3491B83992B5}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{313AEA56-2B75-4937-B67B-9979D79428C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{313AEA56-2B75-4937-B67B-9979D79428C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{313AEA56-2B75-4937-B67B-9979D79428C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{313AEA56-2B75-4937-B67B-9979D79428C9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4629A941-429D-4843-BA72-3491B83992B5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4629A941-429D-4843-BA72-3491B83992B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4629A941-429D-4843-BA72-3491B83992B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4629A941-429D-4843-BA72-3491B83992B5}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9BCA2C1C-6291-4CFE-B98D-42A5985DD407}
+	EndGlobalSection
+EndGlobal
