+++ Healenium.Selenium.Tests/Healenium.Selenium/constants/LocatorType.cs
@@ -1,0 +1,16 @@
+﻿using System;
+namespace Healenium.Selenium.constants
+{
+    public enum LocatorType
+    {
+        Xpath,
+        Css,
+        Id,
+        LinkText,
+        Name,
+        PartialLinkText,
+        Tag,
+        ClassName
+    }
+}
+
