+++ Healenium.Selenium.Tests/Healenium.Selenium/pages/BasePage.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using OpenQA.Selenium;
+
+namespace Healenium.Selenium.pages
+{
+    public class BasePage
+    {
+        protected IWebDriver Driver;
+
+        public BasePage(IWebDriver driver)
+        {
+            Driver = driver;
+        }
+    }
+}
+
