+++ Healenium.Selenium.Tests/Healenium.Selenium/search/Strategy.cs
@@ -1,0 +1,9 @@
+﻿using System;
+namespace Healenium.Selenium.search
+{
+    public interface Strategy
+    {
+        bool DoAction(string selector);
+    }
+}
+
