--- src/Healenuim.Selenium/constants/LocatorType.cs
@@ -1,14 +1,0 @@
-﻿namespace Healenuim.Selenium.constants
-{
-    public enum LocatorType
-    {
-        Xpath,
-        Css,
-        Id,
-        LinkText,
-        Name,
-        PartialLinkText,
-        Tag,
-        ClassName
-    }
-}
