--- src/Healenuim.Selenium/constants/PageUrl.cs
@@ -1,9 +1,0 @@
-﻿namespace Healenuim.Selenium.constants
-{
-    public class PageUrl
-    {
-        public const string TestEnvUrl = "https://elenastepuro.github.io/test_env/index.html";
-        public const string CallbackUrl = "https://mdn.github.io/web-components-examples/life-cycle-callbacks/";
-        public const string MarkupUrl = "http://sha-test-app.herokuapp.com/";
-    }
-}
