--- src/Healenuim.Selenium/pages/BasePage.cs
@@ -1,14 +1,0 @@
-﻿using OpenQA.Selenium;
-
-namespace Healenuim.Selenium.pages
-{
-    public class BasePage
-    {
-        protected IWebDriver Driver;
-
-        public BasePage(IWebDriver driver)
-        {
-            Driver = driver;
-        }
-    }
-}
