--- src/Healenuim.Selenium/search/Strategy.cs
@@ -1,7 +1,0 @@
-﻿namespace Healenuim.Selenium.search
-{
-    public interface Strategy
-    {
-        bool DoAction(string selector);
-    }
-}
