--- src/Healenuim.Selenium/search/locators/ClassNameStrategy.cs
@@ -1,23 +1,0 @@
-﻿using OpenQA.Selenium;
-
-namespace Healenuim.Selenium.search.locators
-{
-    public class ClassNameStrategy : Strategy
-    {
-        private IWebDriver Driver;
-
-        public ClassNameStrategy()
-        {
-        }
-
-        public ClassNameStrategy(IWebDriver driver)
-        {
-            Driver = driver;
-        }
-
-        public bool DoAction(string selector)
-        {
-            return Driver.FindElement(By.ClassName(selector)).Displayed;
-        }
-    }
-}
