--- src/Healenuim.Selenium/search/locators/XPathStrategy.cs
@@ -1,23 +1,0 @@
-﻿using OpenQA.Selenium;
-
-namespace Healenuim.Selenium.search.locators
-{
-    public class XPathStrategy : Strategy
-    {
-        private IWebDriver Driver;
-
-        public XPathStrategy()
-        {
-        }
-
-        public XPathStrategy(IWebDriver driver)
-        {
-            Driver = driver;
-        }
-
-        public bool DoAction(string selector)
-        {
-            return Driver.FindElement(By.XPath(selector)).Displayed;
-        }
-    }
-}
