--- tests/Healenium.Selenium.Tests/packages.config
@@ -1,20 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="App.Metrics.Abstractions" version="4.2.0" targetFramework="net48" />
-  <package id="Autofac" version="6.3.0" targetFramework="net48" />
-  <package id="Autofac.Extensions.DependencyInjection" version="7.2.0" targetFramework="net48" />
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="1.1.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="2.1.0" targetFramework="net48" />
-  <package id="NUnit" version="3.13.2" targetFramework="net48" />
-  <package id="NUnit3TestAdapter" version="4.0.0" targetFramework="net48" />
-  <package id="Selenium.Support" version="4.1.1" targetFramework="net48" />
-  <package id="Selenium.WebDriver" version="4.1.1" targetFramework="net48" />
-  <package id="Selenium.WebDriver.ChromeDriver" version="101.0.4951.4100" targetFramework="net48" />
-  <package id="Selenium.WebDriverBackedSelenium" version="3.141.0" targetFramework="net48" />
-  <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
-  <package id="System.Diagnostics.DiagnosticSource" version="4.7.1" targetFramework="net48" />
-  <package id="System.Memory" version="4.5.4" targetFramework="net48" />
-  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net48" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.3" targetFramework="net48" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.2" targetFramework="net48" />
-</packages>
