--- tests/Healenium.Selenium.Tests/tests/WaitTest.cs
@@ -1,24 +1,0 @@
-﻿using NUnit.Framework;
-
-namespace Healenium.Selenium.Tests.tests
-{
-    public class WaitTest : BaseTest
-    {
-        [Test]
-        [Description("Conditional wait for simple locator")]
-        public void TestConditionWait()
-        {
-            _mainPage.OpenPage()
-                    .ClickTestButton()
-                    .ConfirmAlert();
-
-            _mainPage.GenerateMarkup()
-                    .ClickTestButton() //should be healed
-                    .ConfirmAlert();
-
-            _mainPage.GenerateMarkup()
-                    .ClickTestButtonWaitor(5) //should be healed
-                    .ConfirmAlert();
-        }
-    }
-}
